made all scores templated and add a heuristic to switch between int16_t and int32_t based on read lengths
